move calculate logic back to calculate
